Move the ActionType type to its own file
