+++ packages/ui/src/utils/ActionType.ts
@@ -1,0 +1,1 @@
+export type ActionType = 'default' | 'accent' | 'unshield' | 'destructive';
